Use *.slug instead of category/feed as latter includes relative URL.
The page_name variable includes relative path to the actual page_name
so it won't necessary match the feed file name if they are not in
the same path.

The category.slug and tag.slug variables instead are just a
shorthand for slugified category/tag name, which is what we need.