Add a template for error pages
The template is just a page but add a `<base href="/">` tag to make sure
that resources for the page will always be loaded from the root.

This fixes missing CSS and JS when encountering an error in a
subdirectory of the site (ie.  http://example.org/blog/doesnt_exist)

To use it, add `template: error` to the metadata of the error page